backend/requirements.txt
Flask
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
Werkzeug==2.3.7
python-dotenv==1.0.0
flask-socketio==5.3.4
